+++ pakyow-realtime/lib/pakyow/realtime/websocket.rb
@@ -38,26 +38,13 @@ module Pakyow
       include Pakyow::Application::Helpers::Application
       include Pakyow::Application::Helpers::Connection
 
-      attr_reader :id
+      attr_reader :id, :connection, :logger
 
       def initialize(id, connection)
         @id, @connection, @open = id, connection, false
         @logger = Logger.new(:sock, id: @id[0..7], output: Pakyow.global_logger, level: Pakyow.config.logger.level)
         @server = @connection.app.websocket_server
-
-        response = Async::WebSocket::Adapters::Native.open(@connection.request, handler: Connection) do |socket|
-          @socket = socket
-
-          handle_open
-          while message = socket.read
-            handle_message(message)
-          end
-        rescue EOFError, Protocol::WebSocket::ClosedError
-        ensure
-          @socket&.close; shutdown
-        end
-
-        @connection.__getobj__.instance_variable_set(:@response, response)
+        open
       end
 
       def open?
@@ -96,6 +83,22 @@ module Pakyow
 
       private
 
+      def open
+        response = Async::WebSocket::Adapters::Native.open(@connection.request, handler: Connection) do |socket|
+          @socket = socket
+
+          handle_open
+          while message = socket.read
+            handle_message(message)
+          end
+        rescue EOFError, Protocol::WebSocket::ClosedError
+        ensure
+          @socket&.close; shutdown
+        end
+
+        @connection.__getobj__.instance_variable_set(:@response, response)
+      end
+
       def handle_open
         @server.socket_connect(self)
         @open = true
@@ -105,10 +108,17 @@ module Pakyow
         start_heartbeat
       end
 
-      def handle_message(message)
+      def handle_message(raw)
         @logger.internal {
-          "< " + message
+          "< " + raw
         }
+
+        message = JSON.parse(raw)
+        if handlers = @connection.app.class.__websocket_handlers[message["type"]]
+          handlers.each do |handler|
+            instance_exec(message["payload"], &handler)
+          end
+        end
       end
 
       def trigger_presence(event)